Use spotify uri lookup if search fails
We start with a search lookup of echonest as that gives us a good chance
of success regardless of which track or tracks are listed for the song
on echonest. But when that doesn't work, we now fall back to looking up
the primary spotify uri; we don't loop through all of them, which would
be the most robust route, as that's probably overkill for now.